requirements.txt
python-telegram-bot>=21.0
litellm>=1.40.0
